requirements.txt
asgiref==3.3.4
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Deprecated==1.2.12
Django==3.2.3
django-appconf==1.0.4
django-braces==1.14.0
django-cors-headers==3.7.0
django-cryptography==1.0
django-oauth-toolkit==1.5.0
django-six==1.0.4
djangorestframework==3.12.4
drf-social-oauth2==1.0.9
ecdsa==0.17.0
gunicorn==20.1.0
idna==2.10
install==1.3.4
jwcrypto==0.9
oauthlib==3.1.1
Pillow==8.2.0
psycopg2==2.9.1
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
python-decouple==3.4
python-jose==3.3.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
twilio==7.3.0
urllib3==1.26.5
whitenoise==5.3.0
wrapt==1.12.1
